+++ ezdxf/lldxf/repair.py
@@ -9,13 +9,12 @@
 from __future__ import unicode_literals
 __author__ = "mozman <mozman@gmx.at>"
 
-from datetime import datetime
 from functools import partial
 import logging
 
 from .extendedtags import ExtendedTags
 from .tags import DXFTag, Tags
-from .const import DXFInternalEzdxfError, DXFStructureError, DXFValueError, DXFKeyError
+from .const import DXFInternalEzdxfError, DXFValueError, DXFKeyError
 
 logger = logging.getLogger('ezdxf')
 
@@ -148,7 +147,6 @@ def upgrade_to_ac1015(dwg):
     upgrade_dim_style_table()
     upgrade_objects()
 
-    add_upgrade_comment(dwg, dwg.dxfversion, 'AC1015 (R2000)')
     dwg.dxfversion = 'AC1015'
     dwg.header['$ACADVER'] = 'AC1015'
 
@@ -156,20 +154,9 @@ def upgrade_to_ac1015(dwg):
 def upgrade_to_ac1009(dwg):
     """Upgrade DXF versions prior to AC1009 (R12) to AC1009.
     """
-    add_upgrade_comment(dwg, dwg.dxfversion, 'AC1009 (R12)')
     dwg.dxfversion = 'AC1009'
     dwg.header['$ACADVER'] = 'AC1009'
     # as far I know, nothing else to do
-
-
-def add_upgrade_comment(dwg, from_version, to_version):
-    from .. import VERSION
-    dwg.dxfversion = 'AC1009'
-    dwg.comments.append("DXF version upgrade from {f} to {t} by ezdxf {v} on {dt}".format(
-        f=from_version,
-        t=to_version,
-        v=VERSION,
-        dt=datetime.now().strftime("%Y-%m-%d %H:%M:%S")))
 
 
 def repair_leica_disto_r12(dwg):